Separate SubWikiPageFactory from FileSystemPageFactory.
